Turned the first test into a repeatable, configurable pattern
